api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]
geoalchemy2
pydantic
python-dotenv
